refactor(cli): fold register_v2 into register + add leaf-layering gate (review R5 part 1)
- ARCH-8: merge register_v2.go into register.go. The `_v2` suffix implied a
  V1 sibling that no longer exists after the V1 removal; the onboarding body
  now lives beside its command definition in one cohesive file.
- ARCH-2: add tests/arch/layering_test.go pinning the intra-internal/cli
  dependency order — leaf packages (clictx/prompt/apispec/binder) import no
  sibling command/UX package, the shared cmdcore base never reaches into a
  command tree, and the api (`jentic`) and ctl(cmd) (`jenticctl`) trees never
  import each other. Guards the import-cycle-break clictx exists for, at a
  finer grain than the whole-binary boundary gate.
